fix(pricing-card and downloadPartner): links update
