+++ src/components/strips/Pricing.tsx
@@ -23,6 +23,7 @@ const Pricing = () => {
               amount={plan.amount}
               rowContent={plan.rowContent}
               actionContent={plan.actionContent}
+              link={plan.link}
               isDisabled={plan.isDisabled}
             />
           ))}
