+++ src/data/download-partner.tsx
@@ -1,0 +1,15 @@
+import PlayStoreButton from '~/svg/PlayStoreVector.svg';
+import AppStoreButton from '~/svg/AppStoreVector.svg';
+
+export const downloadButton = [
+  {
+    label: 'Playstore Button',
+    icon: <PlayStoreButton width={144} height={33} />,
+    link: 'https://play.google.com/store/apps/details?id=com.meracars.mera_partners&hl=en',
+  },
+  {
+    label: 'App store button',
+    icon: <AppStoreButton width={144} height={33} />,
+    link: '/',
+  },
+];
